Add terraform annotations to buildkite plan steps
This adds pretty looking annotations to the Buildkite build steps when
some sort of plan is done.

Based on the detailed exitcode, we create a success, info or error style
annotation, and for error or success with changes plans, we embed a
human readable version of the plan output in the annotation.