Move annotation to a function in buildkite.sh
This moves the logic for annotating to the buildkite.sh file, since it
should be decoupled from the plan logic (if toolbox is being used
without buildkite).

This means that we can no-long use the return code of the plan to figure
out what the result was. However, we can use the presence of a plan file
to determine if the plan succeeded, and we can inspect the plan file to
see if there are any changes.