Switch plan annotations to link to changes instead of showing them
The rendering of the intended plan in the plan annotations didn't work
very well. It was hard to read due to presumably issues with how the
rendered output of the terminal-to-html output is handled.

This simplifies things a little, and just links off to the job when a
plan has changes, and the user can find the plan themselves. Not ideal,
but I think a little cleaner for the moment.

We still get the nice property of being able to immediately see the
outcome of the plans at the top of the build.