Handle the case where there are no resources (#12)
If the plan annotate function runs when there are no existing resources,
and no new resources, we get a failure. This accounts for that case in
the jq command, by checking if the array that includes the changes is
null, and returning true. This has the effect of annotating with the
green 'successful plan with no changes' annotation.